+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 - MENU : Déplacement du style dans le fichier dédié
 - STYLE : Refactoring en fichier JS
 - SQUELETTE : Nettoyage du code et des fichiers inutiles
+- STYLE : Ajout d'un constructeur de bannière
 
 
 9.4.1 (2022-12-10)